Refactor observation tiles
