+++ explore/src/main/scala/explore/tabs/ObsTabContents.scala
@@ -3,7 +3,6 @@
 
 package explore.tabs
 
-import cats.data.NonEmptyList
 import cats.effect.IO
 import cats.syntax.all._
 import crystal.ViewF
@@ -13,17 +12,12 @@ import eu.timepit.refined.cats._
 import eu.timepit.refined.types.numeric.NonNegInt
 import eu.timepit.refined.types.string.NonEmptyString
 import explore.Icons
-import explore.common.ConstraintSetObsQueriesGQL.AssignConstraintSetToObs
-import explore.common.ConstraintsQueriesGQL._
 import explore.common.ObsQueries._
-import explore.common.TargetQueriesGQL._
 import explore.common.UserPreferencesQueries._
 import explore.common.UserPreferencesQueriesGQL._
 import explore.components.Tile
 import explore.components.TileController
-import explore.components.graphql.LiveQueryRenderMod
 import explore.components.ui.ExploreStyles
-import explore.constraints.ConstraintsPanel
 import explore.implicits._
 import explore.model.Focused.FocusedObs
 import explore.model._
@@ -32,12 +26,9 @@ import explore.model.layout._
 import explore.model.layout.unsafe._
 import explore.model.reusability._
 import explore.observationtree.ObsList
-import explore.schemas.ObservationDB
-import explore.targeteditor.TargetBody
 import japgolly.scalajs.react.MonocleReact._
 import japgolly.scalajs.react._
 import japgolly.scalajs.react.vdom.html_<^._
-import lucuma.core.model.ConstraintSet
 import lucuma.core.model.Target
 import lucuma.core.model.User
 import lucuma.ui.reusability._
@@ -51,11 +42,8 @@ import react.draggable.Axis
 import react.gridlayout._
 import react.resizable._
 import react.resizeDetector.ResizeDetector
-import react.semanticui.addons.select.Select
-import react.semanticui.addons.select.Select.SelectItem
 import react.semanticui.elements.button.Button
 import react.semanticui.elements.button.Button.ButtonProps
-import react.semanticui.modules.dropdown.Dropdown
 import react.semanticui.sizes._
 
 import scala.concurrent.duration._
@@ -70,10 +58,14 @@ final case class ObsTabContents(
   def isObsSelected: Boolean = focused.get.collect { case Focused.FocusedObs(_) => () }.isDefined
 }
 
-object ObsTabContents {
+object ObsTabTiles {
   val NotesId: NonEmptyString         = "notes"
   val TargetId: NonEmptyString        = "target"
   val ConstraintsId: NonEmptyString   = "constraints"
+  val ConfigurationId: NonEmptyString = "configuration"
+}
+
+object ObsTabContents {
   val NotesMaxHeight: NonNegInt       = 3
   val NotesMinHeight: NonNegInt       = 1
   val TargetMinHeight: NonNegInt      = 12
@@ -87,20 +79,20 @@ object ObsTabContents {
                  y = 0,
                  w = DefaultWidth.value,
                  h = NotesMaxHeight.value,
-                 i = NotesId.value,
+                 i = ObsTabTiles.NotesId.value,
                  isResizable = false
       ),
       LayoutItem(x = 0,
                  y = NotesMaxHeight.value,
                  w = DefaultWidth.value,
                  h = TargetMinHeight.value,
-                 i = TargetId.value
+                 i = ObsTabTiles.TargetId.value
       ),
       LayoutItem(x = 0,
                  y = NotesMaxHeight.value + TargetMinHeight.value,
                  w = DefaultWidth.value,
                  h = ConstraintsMaxHeight.value,
-                 i = ConstraintsId.value
+                 i = ObsTabTiles.ConstraintsId.value
       )
     )
   )
@@ -111,20 +103,20 @@ object ObsTabContents {
                  y = 0,
                  w = DefaultWidth.value,
                  h = NotesMaxHeight.value,
-                 i = NotesId.value,
+                 i = ObsTabTiles.NotesId.value,
                  isResizable = false
       ),
       LayoutItem(x = 0,
                  y = NotesMaxHeight.value,
                  w = DefaultWidth.value,
                  h = TargetMinHeight.value,
-                 i = TargetId.value
+                 i = ObsTabTiles.TargetId.value
       ),
       LayoutItem(x = 0,
                  y = NotesMaxHeight.value + TargetMinHeight.value,
                  w = DefaultWidth.value,
                  h = ConstraintsMaxHeight.value,
-                 i = ConstraintsId.value
+                 i = ObsTabTiles.ConstraintsId.value
       )
     )
   )
@@ -135,20 +127,20 @@ object ObsTabContents {
                  y = 0,
                  w = DefaultWidth.value,
                  h = NotesMaxHeight.value,
-                 i = NotesId.value,
+                 i = ObsTabTiles.NotesId.value,
                  isResizable = false
       ),
       LayoutItem(x = 0,
                  y = NotesMaxHeight.value,
                  w = DefaultWidth.value,
                  h = TargetMinHeight.value,
-                 i = TargetId.value
+                 i = ObsTabTiles.TargetId.value
       ),
       LayoutItem(x = 0,
                  y = NotesMaxHeight.value + TargetMinHeight.value,
                  w = DefaultWidth.value,
                  h = ConstraintsMaxHeight.value,
-                 i = ConstraintsId.value
+                 i = ObsTabTiles.ConstraintsId.value
       )
     )
   )
@@ -205,23 +197,6 @@ object ObsTabContents {
             .runAsyncAndForgetCB
         }.getOrEmpty
       }
-
-    private def constraintsSelectorFn(
-      constraintsSetId: Option[ConstraintSet.Id],
-      obsSummaryOpt:    Option[ObsSummary],
-      observations:     ConstraintsInfo
-    )(implicit ctx:     AppContextIO): VdomNode =
-      Select(
-        value = constraintsSetId.map(_.show).orEmpty, // Set to empty string to clear
-        placeholder = "Select a constraint set",
-        onChange = (a: Dropdown.DropdownProps) =>
-          (obsSummaryOpt, ConstraintSet.Id.parse(a.value.toString)).mapN { (obsId, csId) =>
-            AssignConstraintSetToObs
-              .execute(csId, obsId.id)
-              .runAsyncAndForgetCB *> Callback.log(s"Set to $csId")
-          }.getOrEmpty,
-        options = observations.map(s => new SelectItem(value = s.id.show, text = s.name.value))
-      )
 
     protected def renderFn(
       props:        Props,
@@ -291,91 +266,11 @@ object ObsTabContents {
           props.size.width.getOrElse(0) - treeWidth
         }
 
-      def targetRenderFn(
-        targetId:      Target.Id,
-        renderInTitle: Tile.RenderInTitle,
-        targetOpt:     View[Option[TargetEditQuery.Data.Target]]
-      ): VdomNode =
-        (props.userId.get, targetOpt.get).mapN { case (uid, _) =>
-          TargetBody(uid,
-                     targetId,
-                     targetOpt.zoom(_.get)(f => _.map(f)),
-                     props.searching,
-                     state.zoom(State.options),
-                     renderInTitle
-          )
-        }
-
-      def renderTarget(
-        targetId:      Option[Target.Id],
-        renderInTitle: Tile.RenderInTitle
-      ): VdomNode =
-        targetId
-          .map[VdomNode] { targetId =>
-            LiveQueryRenderMod[ObservationDB,
-                               TargetEditQuery.Data,
-                               Option[TargetEditQuery.Data.Target]
-            ](
-              TargetEditQuery.query(targetId),
-              _.target,
-              NonEmptyList.of(TargetEditSubscription.subscribe[IO](targetId))
-            )((targetRenderFn _).reusable(targetId, renderInTitle))
-              .withKey(s"target-$targetId")
-          }
-          .getOrElse(
-            <.div(ExploreStyles.HVCenter |+| ExploreStyles.EmptyTreeContent,
-                  <.div("No target assigned")
-            )
-          )
-
-      val targetTile =
-        Tile(TargetId, "Target")((renderTarget _).reusable(targetId))
-
-      def renderConstraintsFn(
-        csId:          ConstraintSet.Id,
-        renderInTitle: Tile.RenderInTitle,
-        csOpt:         View[Option[ConstraintSetModel]]
-      ): VdomNode =
-        csOpt.get.map { _ =>
-          <.div(
-            ExploreStyles.ConstraintsObsTile,
-            ConstraintsPanel(csId, csOpt.zoom(_.get)(f => _.map(f)), renderInTitle)
-          )
-        }
-
-      def renderConstraints(
-        constraintsSetId: Option[ConstraintSet.Id],
-        renderInTitle:    Tile.RenderInTitle
-      ): VdomNode =
-        constraintsSetId
-          .map[VdomNode] { csId =>
-            LiveQueryRenderMod[ObservationDB, ConstraintSetQuery.Data, Option[ConstraintSetModel]](
-              ConstraintSetQuery.query(csId),
-              _.constraintSet,
-              NonEmptyList.of(ConstraintSetEditSubscription.subscribe[IO](csId))
-            )((renderConstraintsFn _).reusable(csId, renderInTitle))
-              .withKey(s"constraint-$targetId")
-          }
-          .getOrElse(
-            <.div(ExploreStyles.HVCenter |+| ExploreStyles.EmptyTreeContent,
-                  <.div("No constraints assigned")
-            )
-          )
-
-      val constraintsTile =
-        Tile(ConstraintsId,
-             "Constraints",
-             canMinimize = true,
-             control = ((constraintsSelectorFn _)
-               .reusable(constraintsSetId, obsSummaryOpt, observations.get._1))
-               .some
-        )(
-          (renderConstraints _).reusable(constraintsSetId)
-        )
+      val layouts = ViewF.fromState[IO]($).zoom(State.layouts)
 
       val notesTile =
         Tile(
-          NotesId,
+          ObsTabTiles.NotesId,
           s"Note for Observer",
           backButton.some,
           canMinimize = true
@@ -391,15 +286,22 @@ object ObsTabContents {
           )
         )
 
-      val layouts = ViewF.fromState[IO]($).zoom(State.layouts)
-
       val rightSideRGL =
         TileController(
           props.userId.get,
           coreWidth,
           defaultLayout,
           layouts,
-          List(notesTile, targetTile, constraintsTile)
+          List(
+            notesTile,
+            TargetTile.targetTile(props.userId.get,
+                                  targetId,
+                                  props.searching,
+                                  state.zoom(State.options)
+            ),
+            ConstraintsTile
+              .constraintsTile(constraintsSetId, obsSummaryOpt, observations.get._1)
+          )
         )
 
       val rightSide =
